add authentication, post linked to a user

+++ controllers/post.js
@@ -1,5 +1,6 @@
 import { validationResult } from 'express-validator';
 import { Post } from '../models/post.js';
+import { User } from '../models/user.js';
 
 export const getPosts = (req, res, next) => {
   Post.find()
@@ -47,6 +48,8 @@ export const createPost = (req, res, next) => {
     throw error;
   }
   const { content, imageUrl } = req.body;
+  let postUser;
+  let creator;
 
   if (!imageUrl) {
     return;
@@ -55,7 +58,7 @@ export const createPost = (req, res, next) => {
   Post.create({
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'Lucho' },
+    creator: req.userId,
   })
     .then((post) => {
       if (!post) {
@@ -63,9 +66,27 @@ export const createPost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
       }
+      postUser = post;
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      if (!user) {
+        const error = new Error('Something went wrong!');
+        error.statusCode = 422;
+        throw error;
+      }
+      creator = user;
+      user.posts.push(postUser);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: 'Post created!',
-        post: post,
+        post: postUser,
+        creator: {
+          _id: creator._id,
+          name: creator.name,
+        },
       });
     })
     .catch((err) => {
@@ -91,6 +112,13 @@ export const editPost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
       }
+      
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
+
       post.content = content;
       post.imageUrl = imageUrl;
       return post.save();
@@ -108,16 +136,40 @@ export const editPost = (req, res, next) => {
 
 export const deletePost = (req, res, next) => {
   const postId = req.params.postId;
-  Post.findByIdAndRemove(postId)
+  let postDeleted;
+
+  Post.findById(postId)
+    .then((post) => {
+      if (!post) {
+        const error = new Error('Could not find post.');
+        error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
+      return Post.findByIdAndRemove(postId);
+    })
+
     .then((post) => {
       if (!post) {
         const error = new Error('This posts not exits!');
         error.statusCode = 422;
         throw error;
       }
+      postDeleted = post;
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postDeleted);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({
         message: 'Post deleted successfully!',
-        post: post,
+        post: postDeleted,
       });
     })
     .catch((err) => {
